break up listview
