+++ lib/screens/collective/perspectives/collective_populated_list.dart
@@ -3,6 +3,7 @@ import 'package:junto_beta_mobile/backend/backend.dart';
 import 'package:junto_beta_mobile/screens/collective/bloc/collective_bloc.dart';
 import 'package:junto_beta_mobile/screens/collective/perspectives/get_more_expressions_button.dart';
 import 'package:junto_beta_mobile/widgets/custom_feeds/custom_listview.dart';
+import 'package:junto_beta_mobile/widgets/custom_feeds/single_listview.dart';
 import 'package:provider/provider.dart';
 
 class CollectivePopulatedList extends StatelessWidget {
